Remove all logics from GoodsControllers to Service.

+++ src/main/java/service/GoodsService.java
@@ -6,6 +6,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.util.IllegalFormatException;
 import java.util.List;
 
 @Service
@@ -20,18 +21,48 @@ public class GoodsService {
     }
 
     @Transactional
-    public void addGoods(String name, String description, int quantity, int price) {
-        goodsDAO.addGoods(name, description, quantity, price);
+    public boolean addGoods(String name, String description, String quantity, String price) {
+        try {
+            int qua = Integer.parseInt(quantity);
+            int pri = Integer.parseInt(price);
+            if (qua > 0 && pri > 0 && !name.trim().equals("")
+                    && !description.trim().equals("")) {
+                goodsDAO.addGoods(name,description,qua,pri);
+                return true;
+            } else {
+                return false;
+            }
+        } catch (NumberFormatException | IllegalFormatException e){
+            return false;
+        }
     }
 
     @Transactional
-    public void deleteGoods(int id) {
-        goodsDAO.deleteGoods(id);
+    public void deleteGoods(String button) {
+        List<Goods> list = getAllGoods();
+        for (int i = 0; i< list.size(); i++) {
+            int id = list.get(i).getId();
+            if (button.equals("Delete " + i)) {
+                goodsDAO.deleteGoods(id);
+            }
+        }
     }
 
     @Transactional
-    public void editGoods(int id, String name, String description, int quantity, int price) {
-        goodsDAO.editGoods(id, name, description, quantity, price);
+    public boolean editGoods(String id, String name, String description, String quantity, String price) {
+        List<Goods> list = getAllGoods();
+        int i = Integer.parseInt(id);
+        int qua = Integer.parseInt(quantity);
+        int pri = Integer.parseInt(price);
+        if (i > 0 || pri > 0 || qua > 0){
+            for (Goods goods : list) {
+                if (goods.getId() == i) {
+                    goodsDAO.editGoods(i, name, description, qua, pri);
+                    return true;
+                }
+            }
+        }
+        return false;
     }
 
     @Transactional
